Tried The Mask Based Score Calculation
This tech is not working properly, but it's clear that it's quite resource intensive.

I'd like to give it another shot at a later point, but for now, we'll put it to bed.

Look into Compute Shaders and Command Buffers when you come back Erik - you were close to an epiphany in the Paint Manager class.